added email notification

+++ app/Services/B2B/ProductCategoryService.php
@@ -11,6 +11,7 @@ use App\Models\B2BProduct;
 use App\Models\B2bProductCategory;
 use App\Models\B2bProductSubCategory;
 use App\Trait\HttpResponse;
+use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Support\Str;
 
 class ProductCategoryService
@@ -24,8 +25,9 @@ class ProductCategoryService
                 $url = uploadImage($request, 'image', 'category');
             }
 
-            B2BProductCategory::create([
+            B2bProductCategory::create([
                 'name' => $request->name,
+                'type' => $request->type,
                 'slug' => Str::slug($request->name),
                 'image' => $url['url'] ?? null,
                 'featured' => 1,
@@ -39,7 +41,7 @@ class ProductCategoryService
 
     public function updateCategory($request, $id)
     {
-        $category = B2BProductCategory::findOrFail($id);
+        $category = B2bProductCategory::findOrFail($id);
 
         try {
             if ($request->file('image')) {
@@ -58,9 +60,18 @@ class ProductCategoryService
         }
     }
 
+    public function singleCategory($id)
+    {
+        $category = B2bProductCategory::with(['products', 'subcategory'])
+            ->where('id', $id)
+            ->firstOrFail();
+
+        return $this->success(new B2BCategoryResource($category), 'Category details');
+    }
+
     public function categories()
     {
-        $categories = B2BProductCategory::where('featured', 1)
+        $categories = B2bProductCategory::where('featured', 1)
             ->latest()
             ->take(10)
             ->get();
@@ -71,13 +82,15 @@ class ProductCategoryService
     public function adminCategories($request)
     {
         $search = $request->query('search');
+        $type = $request->query('type');
 
-        $categories = B2BProductCategory::with(['products', 'subcategory'])
+        $categories = B2bProductCategory::with(['products', 'subcategory'])
             ->withCount(['products', 'subcategory'])
-            ->when($search, function ($query, string $search): void {
-                $query->where('name', 'like', '%'.$search.'%');
+            ->when($type, fn ($q) => $q->where('type', $type))
+            ->where(function (Builder $query) use ($search) {
+                $query->where('name', 'LIKE', '%'.$search.'%');
+
             })
-            ->latest()
             ->get();
 
         return $this->success(AdminCategoryResource::collection($categories), 'Categories retrieved successfully');
@@ -85,7 +98,7 @@ class ProductCategoryService
 
     public function createSubCategory($request)
     {
-        $category = B2BProductSubCategory::find($request->category_id);
+        $category = B2bProductCategory::find($request->category_id);
 
         if (! $category) {
             return $this->error(null, 'Not found', 404);
@@ -97,10 +110,11 @@ class ProductCategoryService
                 $url = uploadImage($request, 'image', 'subcategory');
             }
 
-            $category->create([
+            B2bProductSubCategory::create([
                 'name' => $request->name,
                 'slug' => Str::slug($request->name),
                 'image' => $url['url'] ?? null,
+                'category_id' => $request->category_id,
             ]);
 
             return $this->success(null, 'Created successfully', 201);
@@ -122,7 +136,7 @@ class ProductCategoryService
 
     public function featuredStatus($request, $id)
     {
-        $category = B2BProductCategory::findOrFail($id);
+        $category = B2bProductCategory::findOrFail($id);
 
         if ($request->has('featured')) {
             $category->featured = $request->input('featured') ? 1 : 0;
@@ -139,7 +153,7 @@ class ProductCategoryService
 
     public function categoryAnalytic()
     {
-        $categories = B2BProductSubCategory::withCount(['subcategory', 'products'])
+        $categories = B2bProductSubCategory::withCount(['subcategory', 'products'])
             ->get();
 
         $totalActive = $categories->where('status', CategoryStatus::ACTIVE)->count();
@@ -167,7 +181,7 @@ class ProductCategoryService
     {
         $search = $request->query('search');
 
-        $subcats = B2BProductSubCategory::with(['products', 'category'])
+        $subcats = B2bProductSubCategory::with(['products', 'category'])
             ->withCount('products')
             ->when($search, function ($query, string $search): void {
                 $query->where('name', 'like', '%'.$search.'%');
@@ -180,7 +194,7 @@ class ProductCategoryService
 
     public function subStatus($request, $id)
     {
-        $sub = B2BProductSubCategory::findOrFail($id);
+        $sub = B2bProductSubCategory::findOrFail($id);
 
         if ($request->filled('status')) {
             $sub->status = $request->input('status') == 1 ? CategoryStatus::ACTIVE : CategoryStatus::INACTIVE;
@@ -195,6 +209,13 @@ class ProductCategoryService
     {
         $category = B2bProductCategory::findOrFail($id);
 
+        if ($category->products()->exists()) {
+            return $this->error(null, 'Cannot delete category: It contains products.', 422);
+        }
+
+        if ($category->subcategory()->exists()) {
+            return $this->error(null, 'Cannot delete category: It contains subcategories.', 422);
+        }
         $category->delete();
 
         return $this->success(null, 'Deleted successfully');
@@ -202,7 +223,7 @@ class ProductCategoryService
 
     public function deleteSubCategory($id)
     {
-        $subcat = B2BProductSubCategory::findOrFail($id);
+        $subcat = B2bProductSubCategory::findOrFail($id);
 
         $subcat->delete();
